Remove the Unlink=off option
This option means rsyslog won't delete the `/rsyslog/imuxsock` when
it shuts down, but rsyslog always expects to be able to create the
socket when starting (unless it has been told by systemd about an
existing socket), so fails to start if the socket already exists.

By removing this option, rsyslog is able to take care creation and
deletion of the socket itself.

Fixes #1