Changing socket path to /var/run/rsyslog/dev/log
This is so that the socket name can be a little more familiar for
client containers mounting in the parent directory, rather than
mounting the socket directly.

This comes as part of a general effort to support usage of this
image by mounting the parent directory of the socket into the
client container and symlinking from `/dev/log`, rather than
mounting the socket itself, as this can cause issues if containers
don't start in the correct order.  Credit for the symlink approach
goes to https://github.com/helderco/docker-rsyslog.